requirements.txt
setuptools~=75.8.2
numpy~=2.2.4
matplotlib~=3.10.1
pillow~=11.1.0
opencv-python~=4.11.0.86